% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_single_sheet.R
\name{generate_single_sheet}
\alias{generate_single_sheet}
\title{Generate a datasheet for single factor treatment experiments}
\usage{
generate_single_sheet(
  observation_frequency,
  num_records,
  treatments,
  replicates_per_treatment,
  parameters
)
}
\arguments{
\item{observation_frequency}{number of weeks between observations}

\item{num_records}{number of observation records to generate}

\item{treatments}{vector of treatment names}

\item{replicates_per_treatment}{number of replicates for each treatment}

\item{parameters}{vector of dependent variables to measure.}
}
\value{
A dataframe structured for single factor treatment experiments, including unique IDs, weeks, treatment IDs, replicate numbers, and parameter columns initialized with NA values
}
\description{
Generate a datasheet for single factor treatment experiments
}
\examples{
generate_single_sheet(
  observation_frequency = 2,
  num_records = 5,
  treatments = c("0mg/L", "2.5mg/L", "5mg/L", "10mg/L"),
  replicates_per_treatment = 7,
  parameters = c('Number of leaves', 'Number of shoot','Number of roots')

)

}
