% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_single_sheet.R
\name{generate_single_sheet}
\alias{generate_single_sheet}
\title{Generate a datasheet for single factor treatment experiments}
\usage{
generate_single_sheet(
  observation_frequency = NULL,
  num_records = NULL,
  time_unit = c("Weeks", "Days"),
  treatments = NULL,
  replicates_per_treatment = NULL,
  parameters,
  research_title = NULL,
  researcher_mail = NULL,
  researcher_name = NULL,
  date_started = NULL,
  plant_species = NULL,
  explant_type = NULL,
  culture_condition = NULL,
  treatment_desc = NULL,
  replicate_per_treatment = NULL,
  observation_interval = NULL
)
}
\arguments{
\item{observation_frequency}{number of weeks between observations}

\item{num_records}{number of observation records to generate}

\item{treatments}{vector of treatment names}

\item{replicates_per_treatment}{number of replicates for each treatment}

\item{parameters}{vector of dependent variables to measure.}
}
\value{
A dataframe structured for single factor treatment experiments, including unique IDs, weeks, treatment IDs, replicate numbers, and parameter columns initialized with NA values
}
\description{
Generate a datasheet for single factor treatment experiments
}
\examples{

generate_single_sheet(
  observation_frequency = 1,
  num_records = 4,
  time_unit = "Days",
  treatments = c("0mg/L", "2.5mg/L", "5mg/L"),
  replicates_per_treatment = 3,
  parameters = "Number of leaves",
  # Metadata inputs
  research_title = "Effect of Cytokinin Concentration on Yam Node Culture",
  researcher_mail = "israel.example.com",
  researcher_name = "Israel Tetteh",
  date_started = "2025-01-15",
  plant_species = "Dioscorea  rotundata",
  explant_type = "Node",
  culture_condition = "25°C, 16h light / 8h dark",
  treatment_desc = "Cytokinin at 0, 2.5, and 5 mg/L",
  replicate_per_treatment = 3,
  observation_interval = "Weekly"
)

}
